Added a page for editing existing products

+++ shop/produkty/forms.py
@@ -1,9 +1,9 @@
-from wtforms import IntegerField, StringField, BooleanField, TextAreaField, validators, Form
+from wtforms import IntegerField, StringField, BooleanField, TextAreaField, validators, Form, DecimalField
 from flask_wtf.file import FileAllowed, FileField, FileRequired
 
 class dodajProdukty(Form):
     nazwa = StringField('Nazwa',[validators.DataRequired()])
-    cena = IntegerField('Cena',[validators.DataRequired()])
+    cena = DecimalField('Cena',[validators.DataRequired()])
     znizka = IntegerField('Zniżka', default=0)
     ilosc = IntegerField('Ilość',[validators.DataRequired()])
     opis = TextAreaField('Opis',[validators.DataRequired()])
